Revert fdecc11..4af2345
This rolls back to commit fdecc11128596eadc00e7a7d81b856d844396c37.
