+++ cmd/mist/assets/ext/ethereum.js/package.json
@@ -1,42 +1,38 @@
 {
   "name": "ethereum.js",
   "namespace": "ethereum",
-  "version": "0.0.15",
+  "version": "0.0.13",
   "description": "Ethereum Compatible JavaScript API",
   "main": "./index.js",
   "directories": {
     "lib": "./lib"
   },
   "dependencies": {
-    "bignumber.js": ">=2.0.0",
     "ws": "*",
-    "xmlhttprequest": "*"
+    "xmlhttprequest": "*",
+    "bignumber.js": ">=2.0.0"
   },
   "devDependencies": {
     "bower": ">=1.3.0",
     "browserify": ">=6.0",
-    "coveralls": "^2.11.2",
     "del": ">=0.1.1",
     "envify": "^3.0.0",
     "exorcist": "^0.1.6",
     "gulp": ">=3.4.0",
     "gulp-jshint": ">=1.5.0",
     "gulp-rename": ">=1.2.0",
-    "gulp-streamify": "0.0.5",
     "gulp-uglify": ">=1.0.0",
-    "istanbul": "^0.3.5",
     "jshint": ">=2.5.0",
-    "mocha": ">=2.1.0",
-    "mocha-lcov-reporter": "0.0.1",
+    "uglifyify": "^2.6.0",
     "unreachable-branch-transform": "^0.1.0",
-    "vinyl-source-stream": "^1.0.0"
+    "vinyl-source-stream": "^1.0.0",
+    "mocha": ">=2.1.0"
   },
   "scripts": {
     "build": "gulp",
     "watch": "gulp watch",
     "lint": "gulp lint",
-    "test": "mocha",
-    "test-coveralls": "istanbul cover _mocha -- -R spec && cat coverage/lcov.info | coveralls --verbose"
+    "test": "mocha"
   },
   "repository": {
     "type": "git",
