--- cmd/mist/assets/ext/ethereum.js/test/test.utils.js
@@ -1,19 +1,0 @@
-var assert = require('assert');
-
-var methodExists = function (object, method) {
-    it('should have method ' + method + ' implemented', function() {
-        assert.equal('function', typeof object[method], 'method ' + method + ' is not implemented');
-    });
-};
-
-var propertyExists = function (object, property) {
-    it('should have property ' + property + ' implemented', function() {
-        assert.notEqual('undefined', typeof object[property], 'property ' + property + ' is not implemented');
-    });
-};
-
-module.exports = {
-    methodExists: methodExists,
-    propertyExists: propertyExists
-};
-
