+++ tests/files/VMTests/RandomTests/goFail3.json
@@ -26,7 +26,7 @@
                 "address" : "0f572e5295c57f15886f9b263e2f6d2d6c7b5ec6",
                 "bloom" : "00000000000000000000800000000000000000000000000000000880000000000000000000000000000000000000000000000000000000000000000000000000",
                 "data" : "0x",
-                "topic" : [
+                "topics" : [
                 ]
             }
         ],
